Refactor MarkdownImage
