+++ apps/ff-site/src/app/_components/MarkdownContent.tsx
@@ -1,7 +1,5 @@
-import Image from 'next/image'
-
+import { createMarkdownImage } from '@filecoin-foundation/ui/Markdown'
 import { SmartTextLink } from '@filecoin-foundation/ui/TextLink/SmartTextLink'
-import { buildImageSizeProp } from '@filecoin-foundation/utils/buildImageSizeProp'
 import rehypeToc, { type HtmlElementNode } from '@jsdevtools/rehype-toc'
 import * as Sentry from '@sentry/node'
 import ReactMarkdown, { type Components } from 'react-markdown'
@@ -18,40 +16,6 @@ type PluggableList = Parameters<typeof ReactMarkdown>[0]['rehypePlugins']
 export type MarkdownContentProps = {
   children: Parameters<typeof ReactMarkdown>[0]['children']
   addTableOfContents?: boolean
-}
-
-const IMAGE_DIMENSIONS = {
-  containerWidth: 672,
-  aspectRatioHeight: Math.round(672 * (9 / 16)),
-} as const
-
-const MarkdownImage: Components['img'] = ({ src, alt }) => {
-  const commonProps = {
-    quality: 100,
-    width: IMAGE_DIMENSIONS.containerWidth,
-    height: IMAGE_DIMENSIONS.aspectRatioHeight,
-    sizes: buildImageSizeProp({
-      startSize: '100vw',
-      md: `${IMAGE_DIMENSIONS.containerWidth}px`,
-    }),
-  }
-
-  if (!src) {
-    const errorMessage = 'Invalid markdown: image is missing src attribute'
-
-    console.error(errorMessage)
-    Sentry.captureException(new Error(errorMessage))
-
-    return (
-      <Image
-        {...commonProps}
-        src={graphicsData.imageFallback.data}
-        alt={graphicsData.imageFallback.alt}
-      />
-    )
-  }
-
-  return <Image {...commonProps} src={src} alt={alt || ''} />
 }
 
 const MarkdownLink: Components['a'] = ({ href, children }) => {
@@ -72,7 +36,7 @@ const MarkdownLink: Components['a'] = ({ href, children }) => {
 }
 
 const markdownComponents: Components = {
-  img: MarkdownImage,
+  img: createMarkdownImage(graphicsData.imageFallback),
   a: MarkdownLink,
 }
