+++ packages/ui/src/Markdown/index.ts
@@ -1,0 +1,1 @@
+export { createMarkdownImage } from './MarkdownImage'
